+++ src/lib.rs
@@ -1,7 +1,13 @@
 #![doc = include_str!("../README.md")]
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/anergictcell/atg/main/assets/logo_standard.png",
+    html_favicon_url = "https://raw.githubusercontent.com/anergictcell/atg/main/assets/favicon.ico"
+)]
 
 pub mod bed;
 pub mod fasta;
+pub mod genepred;
+pub mod genepredext;
 pub mod gtf;
 pub mod models;
 pub mod refgene;