+++ src/models/transcript.rs
@@ -270,8 +270,8 @@ impl Transcript {
             return vec![];
         }
         let codon = match self.strand {
-            Strand::Minus => Codon::downstream(self, &self.cds_start().unwrap()),
-            Strand::Plus => Codon::upstream(self, &self.cds_end().unwrap()),
+            Strand::Minus => Codon::downstream(self, &self.cds_start().unwrap()), // cannot fail, tx is coding
+            Strand::Plus => Codon::upstream(self, &self.cds_end().unwrap()), // cannot fail, tx is coding
             _ => return vec![],
         };
         if let Ok(res) = codon {
@@ -317,8 +317,8 @@ impl Transcript {
         let mut coords: CoordinateVector = vec![];
         for exon in self.exons() {
             if exon.is_coding() {
-                let cds_start = exon.cds_start().unwrap();
-                let cds_end = exon.cds_end().unwrap();
+                let cds_start = exon.cds_start().unwrap(); // cannot fail, exon is coding
+                let cds_end = exon.cds_end().unwrap(); // cannot fail, exon is coding
                 if cds_start > exon.start() {
                     coords.push((self.chrom(), exon.start(), cds_start - 1))
                 }
@@ -341,10 +341,10 @@ impl Transcript {
         let mut utr = self.utr_coordinates();
         if self.is_coding() {
             if self.forward() {
-                let start = self.cds_start().unwrap();
+                let start = self.cds_start().unwrap(); // cannot fail, exon is coding
                 utr.retain(|coord| coord.2 < start);
             } else {
-                let end = self.cds_end().unwrap();
+                let end = self.cds_end().unwrap(); // cannot fail, exon is coding
                 utr.retain(|coord| coord.1 > end);
             }
         }
@@ -360,10 +360,10 @@ impl Transcript {
         let mut utr = self.utr_coordinates();
         if self.is_coding() {
             if self.forward() {
-                let end = self.cds_end().unwrap();
+                let end = self.cds_end().unwrap(); // cannot fail, exon is coding
                 utr.retain(|coord| coord.1 > end);
             } else {
-                let start = self.cds_start().unwrap();
+                let start = self.cds_start().unwrap(); // cannot fail, exon is coding
                 utr.retain(|coord| coord.2 < start);
             }
             utr
@@ -393,27 +393,27 @@ impl PartialEq for Transcript {
     /// The fields `bin` and `score` are **not** taken into consideration
     /// for equality.
     fn eq(&self, other: &Self) -> bool {
-        if self.name != other.name {
+        if self.name() != other.name() {
             return false;
         };
 
-        if self.chrom != other.chrom {
+        if self.chrom() != other.chrom() {
             return false;
         };
 
-        if self.strand != other.strand {
+        if self.strand() != other.strand() {
             return false;
         };
 
-        if self.gene_symbol != other.gene_symbol {
+        if self.gene() != other.gene() {
             return false;
         };
 
-        if self.cds_start_stat != other.cds_start_stat {
+        if self.cds_start_stat() != other.cds_start_stat() {
             return false;
         };
 
-        if self.cds_end_stat != other.cds_end_stat {
+        if self.cds_end_stat() != other.cds_end_stat() {
             return false;
         };
 
@@ -421,7 +421,7 @@ impl PartialEq for Transcript {
             return false;
         };
 
-        for (exon_a, exon_b) in self.exons.iter().zip(&other.exons) {
+        for (exon_a, exon_b) in self.exons().iter().zip(other.exons()) {
             if exon_a != exon_b {
                 return false;
             }