+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.5
+- Add exon numbers in GTF output
+- Refactor GTF parsing to improve performance
+- Add GenePredExt Reader and Writer
+- Add GenePred Writer
+- Remove chr prefixing during parsing of chromosomes/contigs
+- Improve docs
+- Add more integration tests
+
+## 0.4
 - Support writing Fasta files
 - Support sequence output per feature
 - Improved error handling